% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_labs_3D.R
\name{axis_labs_3D}
\alias{axis_labs_3D}
\title{3D Axis Numbering}
\usage{
axis_labs_3D(mapping = aes(group = 1), data = NULL, geom = "text",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{Arguements passed on to layer.
These are often aesthetics, used to set an
aesthetic to a fixed value, like color = "red" or size = 3.}

\item{theta}{The azimuthal direction in degrees.}

\item{phi}{The colatitude in degrees.}
}
\description{
This function adds 3D axis numbering to ggplot2 plots.
}
