% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_3D.R
\name{stat_3D}
\alias{stat_3D}
\title{Draw 3D Geoms}
\usage{
stat_3D(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{geom}{The geom type to use *ie. "point", "path", "line"*}

\item{...}{Arguements passed on to layer.
These are often aesthetics, used to set an
aesthetic to a fixed value, like color = "red" or size = 3.}

\item{theta}{The azimuthal direction in degrees.}

\item{phi}{The colatitude in degrees.}
}
\description{
This function adds 3D geoms such as points and paths to a ggplot2 plot.
}
