% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_wireframe.R
\name{stat_wireframe}
\alias{stat_wireframe}
\title{Wireframe Plot}
\usage{
stat_wireframe(mapping = NULL, data = NULL, geom = "path",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{Arguements passed on to layer.
These are often aesthetics, used to set an
aesthetic to a fixed value, like color = "red" or size = 3.}

\item{theta}{The azimuthal direction in degrees.}

\item{phi}{The colatitude in degrees.}
}
\description{
This function adds a 3D wireframe to a ggplot2 plot.
}
